Move components folder
