+++ wcag-tool/chrome-extension/src/components/common/ActionButton.jsx
