+++ wcag-tool/chrome-extension/src/components/common/FloatButton.jsx
