+++ wcag-tool/chrome-extension/src/components/common/NavigationButton.jsx
