+++ wcag-tool/chrome-extension/src/components/editor/AnnotationList.jsx
