+++ wcag-tool/chrome-extension/src/components/editor/App.jsx
