+++ wcag-tool/chrome-extension/src/components/editor/EditAnnotationSlider.jsx
@@ -8,7 +8,7 @@ import {
   setEditSliderIsOpen,
   setSelectedAnnotation,
   updateAnnotation,
-} from '../../../services/annotationSlice'
+} from '../../services/annotationSlice'
 
 const EditAnnotationSlider = function () {
   const isOpen = useSelector(selectEditSliderIsOpen)
