+++ wcag-tool/chrome-extension/src/components/editor/NoAnnotation.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useDispatch } from 'react-redux'
-import { setListSliderIsOpen, setSelectElement } from '../../../services/annotationSlice'
+import { setListSliderIsOpen, setSelectElement } from '../../services/annotationSlice'
 import DefaultButton from '../common/DefaultButton'
 
 const NoAnnotation = function () {
