+++ wcag-tool/chrome-extension/src/components/editor/react-sliding-pane.css
