+++ wcag-tool/chrome-extension/src/components/editor/tooltip.css
