+++ wcag-tool/chrome-extension/src/components/popup/navigationButtons.jsx
