+++ wcag-tool/chrome-extension/src/components/popup/searchBar.jsx
