+++ wcag-tool/chrome-extension/src/components/popup/snapshotBody.jsx
